requirements.txt
click==7.1.1
cycler==0.10.0
Cython==0.29.15
dnspython==1.16.0
Flask==1.1.1
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.0
numpy==1.18.1
Pillow==7.0.0
pycocotools==2.0
pymongo==3.10.1
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
Werkzeug==1.0.0
